+++ src/Model/PowerFlow/MiniGrid/MiniGrid.hpp
@@ -58,7 +58,7 @@
  */
 #pragma once
 
-#include <ModelEvaluatorImpl.hpp>
+#include <Model/PowerFlow/ModelEvaluatorImpl.hpp>
 #include <vector>
 
 namespace GridKit
