+++ src/Model/PowerFlow/SystemModel.hpp
@@ -65,7 +65,7 @@
 #include <cassert>
 
 #include <ScalarTraits.hpp>
-#include <ModelEvaluatorImpl.hpp>
+#include <Model/PowerFlow/ModelEvaluatorImpl.hpp>
 
 namespace GridKit
 {
