+++ src/Model/PowerFlow/SystemModelPowerFlow.hpp
@@ -71,7 +71,7 @@
 #include <cassert>
 
 #include <ScalarTraits.hpp>
-#include <ModelEvaluatorImpl.hpp>
+#include <Model/PowerFlow/ModelEvaluatorImpl.hpp>
 
 namespace GridKit
 {
